+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fixed
 - Fixed a bug that blocked arbitrum mainnet wallets from being created
+- Add `deployToken` method to `WalletAddress` and `Wallet` to deploy an ERC20, updated `SmartContract` class to support deployment and fetching contract details
 
 ## [0.5.0] - 2024-09-11