+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Coinbase Node.js SDK Changelog
 
-## Unreleased 
+## Unreleased
 
 ## [0.5.1] - 2024-09-12
 
 ### Fixed
 - Fixed a bug that blocked arbitrum mainnet wallets from being created
 - Add `deployToken` method to `WalletAddress` and `Wallet` to deploy an ERC20, updated `SmartContract` class to support deployment and fetching contract details
+- Add SOL asset support
 
 ## [0.5.0] - 2024-09-11